# src/main/resources/org/semux/api/v1_1_0/swagger.json
{
    "swagger" : "2.0",
    "info" : {
        "description" : "Semux is an experimental high-performance blockchain platform that powers decentralized application.",
        "version" : "1.1.0",
        "title" : "Semux",
        "contact" : {
            "name" : "Semux Foundation",
            "url" : "https://www.semux.org"
        },
        "license" : {
            "name" : "MIT License",
            "url" : "https://opensource.org/licenses/mit-license.php"
        }
    },
    "schemes" : [
        "http"
    ],
    "consumes" : [
        "application/x-www-form-urlencoded"
    ],
    "produces" : [
        "application/json"
    ],
    "host" : "localhost:5171",
    "basePath" : "/v1.0.2",
    "securityDefinitions" : {
        "basicAuth" : {
            "type" : "basic"
        }
    },
    "paths" : {
...
        "/add_node" : {
            "get" : {
                "tags": [
                    "semux"
                ],
                "summary" : "Add node",
                "description" : "Adds a node to node manager.",
                "operationId" : "addNode",
                "produces" : [
                    "application/json"
                ],
                "parameters" : [
                    {
                        "name" : "node",
                        "in" : "query",
                        "description" : "Name of the node in host:port format",
                        "required" : true,
                        "type" : "string"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "successful operation",
                        "schema" : {
                            "$ref" : "#/definitions/AddNodeResponse"
                        }
                    }
                },
                "security" : [
                    {
                        "basicAuth" : [ ]
                    }
                ]
            }
        },
        "/add_to_blacklist" : {
            "get" : {
                "tags": [
                    "semux"
                ],
                "summary" : "Add to blacklist",
                "description" : "Adds an IP address to blacklist.",
                "operationId" : "addToBlacklist",
                "produces" : [
                    "application/json"
                ],
                "parameters" : [
                    {
                        "name" : "ip",
                        "in" : "query",
                        "description" : "IP address",
                        "required" : true,
                        "type" : "string"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "successful operation",
                        "schema" : {
                            "$ref" : "#/definitions/ApiHandlerResponse"
                        }
                    }
                },
                "security" : [
                    {
                        "basicAuth" : [ ]
                    }
                ]
            }
        },
        "/add_to_whitelist" : {
            "get" : {
                "tags": [
                    "semux"
                ],
                "summary" : "Add to whitelist",
                "description" : "Adds an IP address to whitelist.",
                "operationId" : "addToWhitelist",
                "produces" : [
                    "application/json"
                ],
                "parameters" : [
                    {
                        "name" : "ip",
                        "in" : "query",
                        "description" : "IP address",
                        "required" : true,
                        "type" : "string"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "successful operation",
                        "schema" : {
                            "$ref" : "#/definitions/ApiHandlerResponse"
                        }
                    }
                },
                "security" : [
                    {
                        "basicAuth" : [ ]
                    }
                ]
            }
        },
        "/create_account" : {
            "get" : {
                "tags": [
                    "semux"
                ],
                "summary" : "Create account",
                "description" : "Creates a new account.",
                "operationId" : "createAccount",
                "produces" : [
                    "application/json"
                ],
                "parameters" : [
                    {
                        "name" : "name",
                        "in" : "query",
                        "description" : "Assigned alias to the created account.",
                        "required" : false,
                        "type" : "string"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "successful operation",
                        "schema" : {
                            "$ref" : "#/definitions/CreateAccountResponse"
                        }
                    }
                },
                "security" : [
                    {
                        "basicAuth" : [ ]
                    }
                ]
            }
        },
        "/delegate" : {
            "get" : {
                "tags": [
                    "semux"
                ],
                "summary" : "Register delegate",
                "description" : "Registers as a delegate",
                "operationId" : "registerDelegate",
                "produces" : [
                    "application/json"
                ],
                "parameters" : [
                    {
                        "name" : "from",
                        "in" : "query",
                        "description" : "Registering address",
                        "required" : true,
                        "type" : "string"
                    },
                    {
                        "name" : "fee",
                        "in" : "query",
                        "description" : "Transaction fee",
                        "required" : true,
                        "type" : "string",
                        "format" : "int64",
                        "pattern" : "^\\d+$"
                    },
                    {
                        "name" : "data",
                        "in" : "query",
                        "description" : "Delegate name",
                        "required" : true,
                        "type" : "string",
                        "pattern" : "^(0x)?[0-9a-fA-F]+$"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "successful operation",
                        "schema" : {
                            "$ref" : "#/definitions/DoTransactionResponse"
                        }
                    }
                },
                "security" : [
                    {
                        "basicAuth" : [ ]
                    }
                ]
            }
        },
        "/get_account" : {
            "get" : {
                "tags": [
                    "semux"
                ],
                "summary" : "Get account",
                "description" : "Returns an account.",
                "operationId" : "getAccount",
                "produces" : [
                    "application/json"
                ],
                "parameters" : [
                    {
                        "name" : "address",
                        "in" : "query",
                        "description" : "Address of account",
                        "required" : true,
                        "type" : "string"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "successful operation",
                        "schema" : {
                            "$ref" : "#/definitions/GetAccountResponse"
                        }
                    }
                },
                "security" : [
                    {
                        "basicAuth" : [ ]
                    }
                ]
            }
        },
        "/get_account_transactions" : {
            "get" : {
                "tags": [
                    "semux"
                ],
                "summary" : "Get account transactions",
                "description" : "Returns transactions from/to an account.",
                "operationId" : "getAccountTransactions",
                "produces" : [
                    "application/json"
                ],
                "parameters" : [
                    {
                        "name" : "address",
                        "in" : "query",
                        "description" : "Address of account",
                        "required" : true,
                        "type" : "string"
                    },
                    {
                        "name" : "from",
                        "in" : "query",
                        "description" : "Starting range of transactions",
                        "required" : true,
                        "type" : "string"
                    },
                    {
                        "name" : "to",
                        "in" : "query",
                        "description" : "Ending range of transactions",
                        "required" : true,
                        "type" : "string"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "successful operation",
                        "schema" : {
                            "$ref" : "#/definitions/GetAccountTransactionsResponse"
                        }
                    }
                },
                "security" : [
                    {
                        "basicAuth" : [ ]
                    }
                ]
            }
        },
        "/get_block_by_hash" : {
            "get" : {
                "tags": [
                    "semux"
                ],
                "summary" : "Get block by hash",
                "description" : "Returns a block by block hash.",
                "operationId" : "getBlockByHash",
                "produces" : [
                    "application/json"
                ],
                "parameters" : [
                    {
                        "name" : "hash",
                        "in" : "query",
                        "description" : "Hash of block",
                        "required" : true,
                        "type" : "string",
                        "pattern" : "^(0x)?[0-9a-fA-F]+$"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "successful operation",
                        "schema" : {
                            "$ref" : "#/definitions/GetBlockResponse"
                        }
                    }
                },
                "security" : [
                    {
                        "basicAuth" : [ ]
                    }
                ]
            }
        },
        "/get_block_by_number" : {
            "get" : {
                "tags": [
                    "semux"
                ],
                "summary" : "Get block by number",
                "description" : "Returns a block by block number.",
                "operationId" : "getBlockByNumber",
                "produces" : [
                    "application/json"
                ],
                "parameters" : [
                    {
                        "name" : "number",
                        "in" : "query",
                        "description" : "Number of block",
                        "required" : true,
                        "type" : "string",
                        "format" : "int64",
                        "pattern" : "^\\d+$"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "successful operation",
                        "schema" : {
                            "$ref" : "#/definitions/GetBlockResponse"
                        }
                    }
                },
                "security" : [
                    {
                        "basicAuth" : [ ]
                    }
                ]
            }
        },
        "/get_delegate" : {
            "get" : {
                "tags": [
                    "semux"
                ],
                "summary" : "Get a delegate",
                "description" : "Returns a delegate.",
                "operationId" : "getDelegate",
                "produces" : [
                    "application/json"
                ],
                "parameters" : [
                    {
                        "name" : "address",
                        "in" : "query",
                        "description" : "Delegate address",
                        "required" : true,
                        "type" : "string"
                    }
                ],
                "responses" : {
                    "200" : {
                        "description" : "successful operation",
                        "schema" : {
                            "$ref" : "#/definitions/GetDelegateResponse"
                        }
                    }
                },
                "security" : [
                    {
                        "basicAuth" : [ ]
                    }
                ]
            }
        },
        "/get_delegates" : {
            "get" : {
                "tags": [
                    "semux"
                ],
                "summary" : "Get all delegates",
                "description" : "Returns a list of delegates.",
                "operationId" : "getDelegates",
                "produces" : [
                    "application/json"
                ],
                "responses" : {
                    "200" : {
                        "description" : "successful operation",
                        "schema" : {
                            "$ref" : "#/definitions/GetDelegatesResponse"
                        }
                    }
                },
                "security" : [
                    {
                        "basicAuth" : [ ]
                    }
                ]
            }
        },
        "/get_info" : {
            "get" : {
                "tags": [
                    "semux"
                ],
                "summary" : "Get info",
                "description" : "Returns kernel info.",
                "operationId" : "getInfo",
                "produces" : [
                    "application/json"
                ],
                "responses" : {
                    "200" : {
                        "description" : "successful operation",
                        "schema" : {
                            "$ref" : "#/definitions/GetInfoResponse"
                        }
                    }
                },
                "security" : [
                    {
                        "basicAuth" : [ ]
                    }
                ]
            }
        },
        "/get_latest_block" : {
            "get" : {
                "tags": [
                    "semux"
                ],
                "summary" : "Get latest block",
                "description" : "Returns the latest block.",
                "operationId" : "getLatestBlock",
                "produces" : [
                    "application/json"
                ],
                "responses" : {
                    "200" : {
                        "description" : "successful operation",
                        "schema" : {
                            "$ref" : "#/definitions/GetLatestBlockResponse"
                        }
                    }
                },
                "security" : [
                    {
                        "basicAuth" : [ ]
                    }
                ]
            }
        },
        "/get_latest_block_number" : {
            "get" : {
                "tags": [
                    "semux"
                ],
                "summary" : "Get latest block number",
                "description" : "Returns the number of the latest block.",
                "operationId" : "getLatestBlockNumber",
                "produces" : [
                    "application/json"
                ],
                "responses" : {
                    "200" : {
                        "description" : "successful operation",
                        "schema" : {
                            "$ref" : "#/definitions/GetLatestBlockNumberResponse"
                        }
                    }
                },
                "security" : [
                    {
                        "basicAuth" : [ ]
                    }
                ]
            }
        },
...
    },
    "definitions" : {
        "ApiHandlerResponse" : {
            "type" : "object",
            "discriminator" : "ApiHandlerResponse",
            "required" : [
                "success"
            ],
            "properties" : {
                "success" : {
                    "type" : "boolean"
                },
                "message" : {
                    "type" : "string"
                }
            }
        },
...
        "AddNodeResponse" : {
            "type" : "object",
            "allOf" : [
                {
                    "$ref" : "#/definitions/ApiHandlerResponse"
                }
            ]
        },
...
        "CreateAccountResponse" : {
            "type" : "object",
            "allOf" : [
                {
                    "$ref" : "#/definitions/ApiHandlerResponse"
                },
                {
                    "properties" : {
                        "result" : {
                            "type" : "string"
                        }
                    }
                }
            ]
        },
...
    }
}